+++ src/main/java/com/wanmoxing/jishu/bean/UserStudentInfo.java
@@ -10,9 +10,9 @@ public class UserStudentInfo {
 	private String honors;
 	private String contacts;
 	private BigDecimal contactsPrice;
-	private int scoreResponse;
-	private int scoreAttitude;
-	private int scoreProfessional;
+	private BigDecimal scoreResponse;
+	private BigDecimal scoreAttitude;
+	private BigDecimal scoreProfessional;
 	
 	public BigDecimal getContactsPrice() {
 		return contactsPrice;
@@ -50,22 +50,22 @@ public class UserStudentInfo {
 	public void setContacts(String contacts) {
 		this.contacts = contacts;
 	}
-	public int getScoreResponse() {
+	public BigDecimal getScoreResponse() {
 		return scoreResponse;
 	}
-	public void setScoreResponse(int scoreResponse) {
+	public void setScoreResponse(BigDecimal scoreResponse) {
 		this.scoreResponse = scoreResponse;
 	}
-	public int getScoreAttitude() {
+	public BigDecimal getScoreAttitude() {
 		return scoreAttitude;
 	}
-	public void setScoreAttitude(int scoreAttitude) {
+	public void setScoreAttitude(BigDecimal scoreAttitude) {
 		this.scoreAttitude = scoreAttitude;
 	}
-	public int getScoreProfessional() {
+	public BigDecimal getScoreProfessional() {
 		return scoreProfessional;
 	}
-	public void setScoreProfessional(int scoreProfessional) {
+	public void setScoreProfessional(BigDecimal scoreProfessional) {
 		this.scoreProfessional = scoreProfessional;
 	}
